backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlmodel>=0.0.14
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.24.0
websocket-client>=1.6.0
python-multipart>=0.0.6
alembic>=1.11.0
